service/location_subscriber/requirements.txt
paho-mqtt==1.3.1
psycopg2==2.7.3.2
